Round numbers to avoid floating point problems
There are some operations like 0.1 + 0.7 that return weird results like 0.7999999999999999.

This doesn't get along well with the functions that only make sense for integer values (like even or take), so this causes some problems to students.

The aim of this is to reduce the unexpected behaviour from floating point arithmetic by rounding up to 9 digits.

The approach taken was not directly rounding the numbers, but rounding then when they are used in the Ord, Eq and Show functions.
Eq so operations like 0.1 + 0.7 would be equal to 0.8 (even though it would also be equal to 0.7999999999999999)
Ord to keep consistency, because otherwise 0.1 + 0.7 == 0.8 would be true but also 0.1 + 0.7 < 0.8
Show to show the user the rounded value which might be less confusing considering how it's then being used.

This is sure to have some flaws. For example, since we are not actually losing any information, I guess that the digits that are being left out when comparing or ordering due to the rounding could stack by several operations until they actually change the value of the rounded number. However, I'm not sure how often one would get to those cases.