Set other modules to none for library
It seems that in one of the latest versions of hpack, when generating
the cabal file from the package.yaml a module called Path_proyectname
was being added to other-modules.

It usually is not problematic but in our case it was because that was
importing the Prelude when running `stack ghci`.

We set other-modules to [] instead of the recommended solution because
that one implied using the project name but in undercases and that was
hard to implement with the template.